refactor: 로그인 reducer 코드 개선 (#60)
